requirements.txt
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cloudinary==1.28.0
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-MySQL==1.5.2
flask-paginate==2022.1.8
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.1